First commit from source project.
This is a transfer of code from the project `find-ralink`. The purpose
is to turn this into a reusable library, so that multiple command-line
tools can search for wifi adapters that can be used in the IMLS data
collection pilot.

This is an untested first commit. Tests ported over from the original
project; no code changed (substantively). All tests should continue to
pass.