Versioned releases uploading via `gh`
By doing a

make VERSION=v0.1.4 release

we can override the VERSION variable and set the version for a new
release. This will then update the `version.txt` as well as upload
binaries to GH with an appropriate tag and release notes.

(The notes are auto-generated... just an email address, which is
probably redundant?)