The old config code broken out
`config` is now a "library."

This is still heavily co-dependent as a codebase, but... given that we
have multiple command-line tools that need these things, here we are.

The config is needed by many tools because it pulls out paths, and the
auth token work also happens here. So, if we want to log events to the
server, we need the api.data.gov token, and that goes through `auth`.