Drawing pictures works...
Some of the pictures look funny, because we don't really have good
session-based data. However, we can extract quite a few good
"constructed sessions" based on rewriting the data to map to a day.

Broke the data structure out of the data mangler so it could be reused
in the waterfall plotter.