Trying a new ticker.
This may, or may not, drift as much as the original ticker.

It's not clear. It is also not critical; we get data, and useful data,
either way. I'd just rather it was closer to "dead on" every second.

(This may have been a factor of how long our HTTPS POSTs were taking,
though... which would have blocked the process network.)