Added hash cache to mfg db lookup
That lookup is painful. In part, because we're doing (possibly) three
lookups for every MAC address: one of length 14, 8, and *mumble*.

In the testing case, I'm randomly generating MAC addresses, and they're
not being pulled from a valid set of MFG prefixes. (Having deterministic
testing would be nice...)

However, this means we're seeing everything as the "unknown"
manufacturer. And, most of the testing is therefore exercising the
slowest path in the mfg lookup.

The existing unit tests pass (which use real mfgs), but I'd like to do a
bit more with this change.