Fixing the PostJSON error.
I don't like this. However, we're out of time on this phase.

We need a global constant for the storage mode, and need to decide if
we're really going to live this way... but, for now, we're comparing
strings from a YAML file...

This makes sure we don't try and do a PostJSON when we're running in
`storagemode` of "sqlite".