logging added to csp, improvement
Added an input channel to ParDelta. It was treating the 0th element of a
channel array as the input... this was a poor choice. (It worked, but it
was something that could be easily screwed up later...)

I broke out an input channel.

This involved a TLP change in main(). Minor.

Updated _test so the TLP was correct. Also broke out `www` dir for
testing.

The whole network runs again in accelerated test. Confident in changes.