SQLite pushed down.
Pushed some tags down to the struct definitions, which allowed some
abstractions to also push down.

It should be easier to work with structures and DBs now. Also, SQL is no
longer being written in the code, but instead being built in the
`models` section of the main application. That SQL is built based on the
struct definitions/field definitions that are defined in one place.