Interim commit on some restructuring
We are in danger of some circular deps on `config`. I'm pulling a few
things out. The underlying structs (Duration and WifiEvent) are no
longer buried in `analysis`, and are in their own `structs` "library."

`state` will get cleaned up as well (was `tempdb`). It has now become a
set of abstractions for managing our state over the SQLite DBs. I'm
going to take that one step further, and have a `DurationsDB` and a
`WifiEventsDB` abstraction, because we have a lot of DB code that has
leaked upward into the TLP. This required mutex management that led to
nil pointers/crashes that I don't like.

So, doing a quick refactor.