Reworked DB layer.
The tests run (e.g. the TLP tests in session-counter_test.go run), but
no durations are generated. Unsure why at this point. However, queues
and lists seem to be working correctly, logging is working...

This eliminates the Open() and Close() dance that was going on before,
and cleans up config and logger handling significantly.