A simpler wireshark interface...
This is the beginnings of the simplification of how we track devices.

It has some state that would need to be reset (the "macMap").

The essence of this commit is to give us something to talk about/work
with as we think about how to simplify the MAC monitoring portion of the
code. Right now, it is multiple parallel processes, and it's complex.
This simplifies it a great deal (one process?).

To be discussed.