Mocked in some session-counter code
This brings things another step closer to "completing" a top level for
the simplified code.

* Every minute, read data from the shark.
* Every two minutes, log the result.

There's some mocking going on, which will go away shortly. This lets a
Mac/non-RPi run the code (for the moment). And, ProcessData is not done.

When done, it will essentially replace many files/functions, and
simplify the codebase greatly. It will turn EphemeralDurations into
Duration structures, and be done.