Simplifying the abstractions: list
The list abstraction was implemented on top of an SQLite database.

This was... in a world where we thought we might want to recover from
power outages. It added a large amount of complexity.

If we reboot, the world starts over. Nothing is retained.

The list abstraction is now built on top of a... golang slice.