Simplifying/viperifying config
We could use viper. It gives us a singleton config, and the ability to
do env variables and a host of Good Things™.

This will be a breaking change, but I claim it is simpler in the long
run. An SQLite-based configuration file was... not one of my finer
moments. I did it because we had to have SQLite. That is now a liability
for future-proofing/portability.