Need to remove things from memory...
Previously, we would leave things in the DB after being sent, because it
was a local store.

Now, we need to flush memory after sending. We can still cache (in case
we lose network, but not power), but we should remove the data from the
local memory table after it is successfully sent.