Removed code, internal...
Fixed AsMap() so that it mangles a Duration into a JSON object for
posting.

Removed remaining calls to `internal` libraries. Now, everything except
for wifi finding is in this codebase.

Removed some files that were lingering and not necessary anymore.

`api` had nothing left in it; it will be replaced by `resty`.