Deterministic testing in session-counter
This adds a deterministic test, so the same MACs show up in the same
order every minute, and the duration they are present for is constant as
well.

The code for tracking devices (ephemeral) still had a 2h forget window.
This has been removed. It can be added back; it was left commented out.
To be discussed.

The reason... I was chasing determinism. However, this was not a factor.

Note we do not filter devices if they are present more than our target
number of minutes. I don't know where that filtering happens right now,
if it happens at all.